+++ RxUtils/Classes/RxSwift/RxSwift+Deffered.swift
@@ -19,7 +19,7 @@ public extension ObservableType {
     /// - seealso: [startWith operator on reactivex.io](http://reactivex.io/documentation/operators/startwith.html)
     /// - parameter element: Element to prepend to the specified sequence on subscription or `nil`.
     /// - returns: The source sequence prepended with the specified element if it is not `nil`.
-    func startWithDeferred(_ element: @escaping @autoclosure () -> E?) -> Observable<E> {
+    func startWithDeferred(_ element: @escaping @autoclosure () -> Element?) -> Observable<Element> {
         return Observable.create {
             if let element = element() {
                 return self